feat: rename service
