--- server/src/services/lbajob.service.types.ts
@@ -1,7 +1,0 @@
-import { IRecruiter } from "shared"
-
-export interface ILbaJobEsResult {
-  _source: IRecruiter
-  sort?: number[]
-  _id?: string
-}
